+++ data/__init__.py
@@ -1,3 +1,4 @@
 from pre_process import pre_process
+from iterdataset import TextDataset
 
-__all__ = ["pre_process"]
+__all__ = ["pre_process", "TextDataset"]
